+++ source/core/Sprite/SpriteSheet.h
@@ -16,4 +16,4 @@ typedef struct {
 } SpriteSheet;
 
 void SpriteSheet_init(SpriteSheet* sheet, const void* source, SpriteSize sprite_size, SpriteColorFormat format);
-SpriteData SpriteSheet_getSpriteData(SpriteSheet* sheet, u32 index);
+SpriteData SpriteSheet_getSpriteData(SpriteSheet* sheet, u16 index);
